app/requirements.txt
pandas
numpy
tensorflow
scikit-learn
scikeras